+++ crates/stc_ts_type_checker/tests/tsc.rs
@@ -29,7 +29,7 @@ use stc_ts_builtin_types::Lib;
 use stc_ts_env::{Env, ModuleConfig, Rule};
 use stc_ts_file_analyzer::env::EnvFactory;
 use stc_ts_module_loader::resolvers::node::NodeResolver;
-use stc_ts_testing::conformance::TestSpec;
+use stc_ts_testing::conformance::{parse_test, TestSpec};
 use stc_ts_type_checker::{loader::ModuleLoader, Checker};
 use stc_ts_utils::StcComments;
 use swc_common::{
@@ -321,250 +321,6 @@ fn load_expected_errors(ts_file: &Path, spec: Option<&TestSpec>) -> (String, Vec
     )
 }
 
-fn parse_targets(s: &str) -> Vec<(String, EsVersion)> {
-    fn parse_target_inner(s: &str) -> Vec<EsVersion> {
-        match s {
-            "es3" => return vec![EsVersion::Es3],
-            "es5" => return vec![EsVersion::Es5],
-            "es2015" => return vec![EsVersion::Es2015],
-            "es6" => return vec![EsVersion::Es2015],
-            "es2016" => return vec![EsVersion::Es2016],
-            "es2017" => return vec![EsVersion::Es2017],
-            "es2018" => return vec![EsVersion::Es2018],
-            "es2019" => return vec![EsVersion::Es2019],
-            "es2020" => return vec![EsVersion::Es2020],
-            "es2021" => return vec![EsVersion::Es2021],
-            "es2022" => return vec![EsVersion::Es2022],
-            // TODO(upstream): enable es2023
-            // "es2023" => return vec![EsVersion::Es2023],
-            "esnext" => return vec![EsVersion::EsNext],
-            _ => {}
-        }
-        if !s.contains(',') {
-            panic!("failed to parse `{}` as targets", s)
-        }
-        s.split(',').map(|s| s.trim()).flat_map(parse_target_inner).collect()
-    }
-
-    if !s.contains(',') {
-        return vec![(s.into(), parse_target_inner(s)[0])];
-    }
-    s.split(',').map(|s| s.trim()).flat_map(parse_targets).collect()
-}
-
-fn parse_test(file_name: &Path) -> Vec<TestSpec> {
-    let mut err_shift_n = 0;
-    let mut first_stmt_line = 0;
-
-    let fname = file_name.to_string_lossy();
-    ::testing::run_test(false, |cm, handler| {
-        cm.new_source_file(FileName::Anon, "".into());
-
-        let fm = cm.load_file(file_name).expect("failed to read file");
-
-        // We parse files twice. At first, we read comments and detect
-        // configurations for following parse.
-
-        let comments = StcComments::default();
-
-        let mut parser = Parser::new(
-            Syntax::Typescript(TsConfig {
-                tsx: fname.contains("tsx"),
-                ..Default::default()
-            }),
-            SourceFileInput::from(&*fm),
-            Some(&comments),
-        );
-        let mut targets = vec![("".into(), EsVersion::default(), false)];
-
-        let program = parser.parse_program().map_err(|e| {
-            e.into_diagnostic(handler).emit();
-        })?;
-
-        for line in fm.src.lines() {
-            if line.is_empty() {
-                err_shift_n += 1;
-            } else {
-                break;
-            }
-        }
-
-        match &program {
-            Program::Module(v) => {
-                if !v.body.is_empty() {
-                    first_stmt_line = cm.lookup_line(v.body[0].span().lo).unwrap().line;
-                }
-            }
-            Program::Script(v) => {
-                if !v.body.is_empty() {
-                    first_stmt_line = cm.lookup_line(v.body[0].span().lo).unwrap().line;
-                }
-            }
-        }
-
-        let mut libs = vec![Lib::Es5, Lib::Dom];
-        let mut rule = Rule {
-            allow_unreachable_code: false,
-            ..Default::default()
-        };
-        let mut module_config = ModuleConfig::None;
-        let ts_config = TsConfig::default();
-
-        let mut had_comment = false;
-
-        let span = program.span();
-        let comments = comments.leading.get(&span.lo());
-        if let Some(ref comments) = comments {
-            let directive_start = comments.iter().position(|cmt| cmt.text.trim().starts_with('@')).unwrap_or(0);
-            let cmt_start_line = if directive_start == 0 {
-                0
-            } else {
-                comments
-                    .iter()
-                    .find(|cmt| cmt.text.trim().starts_with('@'))
-                    .map(|cmt| cm.lookup_char_pos(cmt.span.hi).line)
-                    .unwrap_or(0)
-            };
-
-            for cmt in comments.iter().skip(directive_start) {
-                let s = cmt.text.trim();
-                if !s.starts_with('@') {
-                    if had_comment {
-                        err_shift_n = cm.lookup_char_pos(cmt.span.hi).line - 1 - cmt_start_line;
-                        break;
-                    }
-                    continue;
-                }
-                had_comment = true;
-                err_shift_n = cm.lookup_char_pos(cmt.span.hi + BytePos(1)).line - cmt_start_line;
-                let s = &s[1..]; // '@'
-
-                if s.starts_with("target:") || s.starts_with("Target:") {
-                    let s = s["target:".len()..].trim().to_lowercase();
-                    targets = parse_targets(&s).into_iter().map(|v| (v.0, v.1, true)).collect();
-                } else if s.starts_with("strict:") {
-                    let strict = s["strict:".len()..].trim().parse().unwrap();
-                    rule.no_implicit_any = strict;
-                    rule.no_implicit_this = strict;
-                    rule.always_strict = strict;
-                    rule.strict_null_checks = strict;
-                    rule.strict_function_types = strict;
-                } else if s.starts_with("noLib:") {
-                    let v = s["noLib:".len()..].trim().parse().unwrap();
-                    if v {
-                        libs = vec![];
-                    }
-                } else if s.to_lowercase().starts_with("noimplicitany:") {
-                    let v = s["noImplicitAny:".len()..].trim().parse().unwrap();
-                    rule.no_implicit_any = v;
-                } else if s.starts_with("noImplicitReturns:") {
-                    let v = s["noImplicitReturns:".len()..].trim().parse().unwrap();
-                    rule.no_implicit_returns = v;
-                } else if s.starts_with("declaration") {
-                } else if s.starts_with("stripInternal:") {
-                    // TODO(kdy1): Handle
-                } else if s.starts_with("traceResolution") {
-                    // no-op
-                } else if s.starts_with("allowUnusedLabels:") {
-                    let v = s["allowUnusedLabels:".len()..].trim().parse().unwrap();
-                    rule.allow_unused_labels = v;
-                } else if s.starts_with("noEmitHelpers") {
-                    // TODO
-                } else if s.starts_with("downlevelIteration:") {
-                    // TODO
-                } else if s.starts_with("sourceMap:") || s.starts_with("sourcemap:") {
-                    // TODO
-                } else if s.starts_with("isolatedModules:") {
-                    // TODO
-                } else if s.starts_with("lib:") {
-                    let s = s["lib:".len()..].trim();
-                    let mut ls = HashSet::<_>::default();
-                    for v in s.split(',') {
-                        ls.extend(Lib::load(&v.to_lowercase().replace("es6", "es2015")))
-                    }
-                    libs = ls.into_iter().collect()
-                } else if s.starts_with("allowUnreachableCode:") {
-                    let v = s["allowUnreachableCode:".len()..].trim().parse().unwrap();
-                    rule.allow_unreachable_code = v;
-                } else if s.starts_with("strictNullChecks:") {
-                    let v = s["strictNullChecks:".len()..].trim().parse().unwrap();
-                    rule.strict_null_checks = v;
-                } else if s.starts_with("noImplicitThis:") {
-                    let v = s["noImplicitThis:".len()..].trim().parse().unwrap();
-                    rule.no_implicit_this = v;
-                } else if s.starts_with("skipDefaultLibCheck") {
-                    // TODO
-                } else if s.starts_with("suppressImplicitAnyIndexErrors:") {
-                    // TODO
-                    let v = s["suppressImplicitAnyIndexErrors:".len()..].trim().parse().unwrap();
-                    rule.suppress_implicit_any_index_errors = v;
-                } else if s.starts_with("module:") {
-                    let v = s["module:".len()..].trim().to_lowercase().parse().unwrap();
-                    module_config = v;
-                } else if s.to_lowercase().starts_with("notypesandsymbols") {
-                    // Ignored as we don't generate them.
-                } else if s.to_lowercase().starts_with("usedefineforclassfields") {
-                    rule.use_define_property_for_class_fields = true;
-                } else if s.to_lowercase().starts_with("jsx") {
-                    rule.jsx = s["jsx:".len()..].trim().to_lowercase().parse().unwrap();
-                } else if s.to_lowercase().starts_with("noemit") || s.to_lowercase().starts_with("preserveconstenums") {
-                    // Ignored as we only checks type.
-                } else if s.starts_with("strict") {
-                    let strict = true;
-                    rule.no_implicit_any = strict;
-                    rule.no_implicit_this = strict;
-                    rule.always_strict = strict;
-                    rule.strict_null_checks = strict;
-                    rule.strict_function_types = strict;
-                } else {
-                    panic!("Comment is not handled: {}", s);
-                }
-            }
-        }
-
-        libs.sort();
-
-        err_shift_n = err_shift_n.min(first_stmt_line);
-        dbg!(err_shift_n);
-
-        Ok(targets
-            .into_iter()
-            .map(|(raw_target, target, specified)| {
-                let libs = if specified && libs == vec![Lib::Es5, Lib::Dom] {
-                    match target {
-                        EsVersion::Es3 | EsVersion::Es5 => vec![Lib::Es5, Lib::Dom],
-                        EsVersion::Es2015 => Lib::load("es2015.full"),
-                        EsVersion::Es2016 => Lib::load("es2016.full"),
-                        EsVersion::Es2017 => Lib::load("es2017.full"),
-                        EsVersion::Es2018 => Lib::load("es2018.full"),
-                        EsVersion::Es2019 => Lib::load("es2019.full"),
-                        EsVersion::Es2021 => Lib::load("es2021.full"),
-                        EsVersion::Es2022 => Lib::load("es2022.full"),
-                        // TODO(upstream): enable es2023
-                        // EsVersion::Es2023 => Lib::load("es2023.full"),
-                        _ => Lib::load("es2022.full"),
-                    }
-                } else if specified {
-                    libs_with_deps(&libs)
-                } else {
-                    libs.clone()
-                };
-
-                TestSpec {
-                    err_shift_n,
-                    libs,
-                    rule,
-                    ts_config,
-                    target,
-                    raw_target,
-                    module_config,
-                }
-            })
-            .collect())
-    })
-    .unwrap()
-}
-
 fn do_test(file_name: &Path, spec: TestSpec, use_target: bool) -> Result<(), StdErr> {
     let (file_stem, mut expected_errors) = load_expected_errors(file_name, if use_target { Some(&spec) } else { None });
     expected_errors.sort();
@@ -820,111 +576,4 @@ impl Fold for Spanner {
     fn fold_span(&mut self, _: Span) -> Span {
         self.span
     }
-}
-
-fn libs_with_deps(libs: &[Lib]) -> Vec<Lib> {
-    fn add(libs: &mut Vec<Lib>, l: Lib) {
-        if libs.contains(&l) {
-            return;
-        }
-        libs.push(l);
-
-        match l {
-            Lib::Es5 | Lib::Es5Full => {}
-
-            Lib::Es2015Collection
-            | Lib::Es2015Core
-            | Lib::Es2015
-            | Lib::Es2015Full
-            | Lib::Es2015Generator
-            | Lib::Es2015Iterable
-            | Lib::Es2015Promise
-            | Lib::Es2015Proxy
-            | Lib::Es2015Reflect
-            | Lib::Es2015Symbol
-            | Lib::Es2015SymbolWellknown => add(libs, Lib::Es2015Full),
-
-            Lib::Es2016ArrayInclude | Lib::Es2016 | Lib::Es2016Full => add(libs, Lib::Es2016Full),
-            Lib::Es2017
-            | Lib::Es2017Full
-            | Lib::Es2017Intl
-            | Lib::Es2017Object
-            | Lib::Es2017Sharedmemory
-            | Lib::Es2017String
-            | Lib::Es2017Typedarrays => add(libs, Lib::Es2017Full),
-
-            Lib::Es2018Asyncgenerator
-            | Lib::Es2018Asynciterable
-            | Lib::Es2018
-            | Lib::Es2018Full
-            | Lib::Es2018Intl
-            | Lib::Es2018Promise
-            | Lib::Es2018Regexp => add(libs, Lib::Es2018Full),
-
-            Lib::Es2019Array
-            | Lib::Es2019
-            | Lib::Es2019Full
-            | Lib::Es2019Object
-            | Lib::Es2019String
-            | Lib::Es2019Symbol
-            | Lib::Es2019Intl => add(libs, Lib::Es2019Full),
-
-            Lib::Es2020Bigint
-            | Lib::Es2020
-            | Lib::Es2020Full
-            | Lib::Es2020Intl
-            | Lib::Es2020Promise
-            | Lib::Es2020Sharedmemory
-            | Lib::Es2020String
-            | Lib::Es2020SymbolWellknown
-            | Lib::Es2020Date
-            | Lib::Es2020Number => add(libs, Lib::Es2020Full),
-
-            Lib::Es2021 | Lib::Es2021Full | Lib::Es2021Weakref | Lib::Es2021Intl | Lib::Es2021Promise | Lib::Es2021String => {
-                add(libs, Lib::Es2021Full)
-            }
-
-            Lib::Es2022
-            | Lib::Es2022Array
-            | Lib::Es2022Error
-            | Lib::Es2022Object
-            | Lib::Es2022Full
-            | Lib::Es2022Intl
-            | Lib::Es2022Sharedmemory
-            | Lib::Es2022String
-            | Lib::Es2022Regexp => add(libs, Lib::Es2022Full),
-
-            Lib::Es2023 | Lib::Es2023Array | Lib::Es2023Full => add(libs, Lib::Es2023Full),
-
-            Lib::Esnext | Lib::EsnextFull | Lib::EsnextIntl | Lib::EsnextPromise | Lib::EsnextString | Lib::EsnextWeakref => {
-                add(libs, Lib::Es2022Full)
-            }
-
-            Lib::Dom
-            | Lib::DomIterable
-            | Lib::DomIterableGenerated
-            | Lib::Header
-            | Lib::Scripthost
-            | Lib::WebworkerGenerated
-            | Lib::WebworkerImportscripts
-            | Lib::WebworkerIterableGenerated
-            | Lib::Decorators
-            | Lib::DecoratorsLegacy => {}
-        }
-
-        for l in l.deps() {
-            add(libs, l);
-        }
-    }
-
-    let mut all = vec![];
-
-    for lib in libs {
-        add(&mut all, *lib);
-    }
-
-    all.sort();
-    all.dedup();
-
-    all
 }
